fix: deprecated error

+++ src/utils/randaug.py
@@ -30,7 +30,7 @@ class RandAugment():
             "translateY": np.linspace(0, 0.2, 10),
             "rotate": np.linspace(0, 360, 10),
             "color": np.linspace(0.0, 0.9, 10),
-            "posterize": np.round(np.linspace(8, 4, 10), 0).astype(np.int),
+            "posterize": np.round(np.linspace(8, 4, 10), 0).astype(np.int32),
             "solarize": np.linspace(256, 231, 10),
             "contrast": np.linspace(0.0, 0.5, 10),
             "sharpness": np.linspace(0.0, 0.9, 10),
